Improved rendering code Adjust legend so it changes numbers with gene/reads threshold Some improvemends in the DEA Fixed some bugs with the thresholds Removed async rendering data computation Restored OpenGLFunctions for compatibility reasons Replaced table in DEA to fix some bugs Fixed a bug in selections table in the sorting of columns Added search box for DEA table Fixed a bug in the DEA threholds Fixed a bug in the minimum value for genes threshold

+++ src/dataModel/Feature.cpp
@@ -22,7 +22,7 @@ Feature::Feature()
 
 }
 
-Feature::Feature(QString barcode, QString gene, int x, int y, int hits)
+Feature::Feature(QString barcode, QString gene, double x, double y, int hits)
     : m_barcode(barcode),
       m_gene(gene),
       m_annotation(),
